Initial Commit
My Second School assignment for ISO development. In this assignment we
made a parallax scrolling background to give the illusion of depth.

There may be an issue with the old framework, last time I opened it on
my Mac it didn't run. Once I get my Mac up and running again I'll update
it with a fix.